Penultimos ajustes acerca do melhoramento do sistema

+++ frontend/src/componentes/usuario/cards/cardReservas/reservas.jsx
@@ -5,10 +5,14 @@ import EstadoCard from "./estado/estado";
 import { motion } from "framer-motion";
 import { useState, useEffect } from "react";
 import api from "../../../service/api/api";
+import Toast from "../../stylenotificacao/toast";
+
+
 
 function CardReservas() {
   const [reservas, setReservas] = useState([]);
   const [emprestimos, setEmprestimos] = useState([]);
+  const [toast, setToast] = useState(null);
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -33,7 +37,7 @@ function CardReservas() {
 
   // 📌 Filtrar reservas e empréstimos
   const livrosReservado = reservas.filter(r =>
-    r.estado_label === "Reservado" || r.estado_label === "Pendente" || r.estado_label === "Em uso"
+    r.estado_label === "Reservado" || r.estado_label === "Pendente" || r.estado_label === "Em Uso"
   );
   const livrosEmprestimo = emprestimos.filter(e =>
     e.acoes === "ativo" || e.acoes === "atrasado"
@@ -41,14 +45,29 @@ function CardReservas() {
 
   // 🗑 Cancelar Reserva
   const handleDeletarReserva = async (reserva) => {
+    if(reserva.estado === "em_uso") {
+      return (
+        setToast({
+          message: "Impossível cancelar reservas em uso!",
+          type: "error",
+        })
+      );
+    }
+
     try {
       await api.delete(`/livros/reservas/${reserva.id}/`);
       setReservas(prev => prev.filter(r => r.id !== reserva.id));
     } catch (error) {
-      if (error.response?.status === 401) navigate("/login");
-      else alert("Erro ao cancelar a reserva: " + JSON.stringify(error.response?.data || error));
+      
+      setToast({
+        message: "Erro de ligação ao servidor. Tente novamente.",
+        type: "error",
+      });
+
     }
   };
+
+  // Em Uso
 
   return (
     <motion.div
@@ -107,13 +126,23 @@ function CardReservas() {
               <Link to={`/detalhes/${reserva.livro_id}`} className="bg-[#F86417] text-white px-4 py-2 rounded-lg cursor-pointer">
                 Ver Detalhes
               </Link>
-              <button onClick={() => handleDeletarReserva(reserva)} className="py-2 px-5 text-black/70 border border-black/17 rounded-lg cursor-pointer">
-                Cancelar Reserva
-              </button>
+              {reserva.estado !== "em_uso" && (
+                <button onClick={() => handleDeletarReserva(reserva)} className="py-2 px-5 text-black/70 border border-black/17 rounded-lg cursor-pointer">
+                  Cancelar Reserva
+                </button>
+              )}
             </div>
           </div>
         </div>
       ))}
+
+      {toast && (
+        <Toast
+          message={toast.message}
+          type={toast.type}
+          onClose={() => setToast(null)}
+        />
+      )}
     </motion.div>
   );
 }
